azure-functions/requirements.txt
sqlalchemy>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
cryptography>=40.0.0,<42.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
passlib[bcrypt]>=1.7.0,<2.0.0
pymysql>=1.0.0,<2.0.0
azure-functions>=1.11.0,<2.0.0
